Added | Delete Doctor Service: Enums of Status hence updated relevant services , Department Id in DoctorDto (reference), Custom Queries for Doctor

+++ src/main/java/com/yt/jpa/hospitalManagement/repository/DepartmentRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DepartmentRepository extends JpaRepository<Department,Integer> {
+public interface DepartmentRepository extends JpaRepository<Department,Long> {
+
 }
